Fixed some integer type conversion warnings.
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@6130 3db68df2-c116-0410-a063-a993310a9797

+++ source/lib/file/archive/archive_zip.cpp
@@ -523,7 +523,7 @@ public:
 		ECDR* ecdr = (ECDR*)pool_alloc(&m_cdfhPool, sizeof(ECDR));
 		if(!ecdr)
 			throw std::bad_alloc();
-		const size_t cd_ofs = m_fileSize;
+		const off_t cd_ofs = m_fileSize;
 		ecdr->Init(m_numEntries, cd_ofs, cd_size);
 
 		m_unalignedWriter->Append(m_cdfhPool.da.base, cd_size+sizeof(ECDR));
